requirements.txt
paho-mqtt>=1.6.0
google-generativeai>=0.3.0
jinja2>=3.1.0
python-dotenv>=1.0.0
scapy>=2.5.0
